[Popover] Fix regression on disabling focus management on hover
